Split credential provider tests
